`_setCollection` overwrites previous data
Previously, `_setCollection` assumed that a collection with data in it
should not be overwritten. Change this behavior to assume that when
`_setCollection` is given data to pass to the collection, we want to
overwrite any pre-existing data.

This is especially useful if a previous router has created the
collection, but entered no or bad data.